% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeeklyMonthlyTrend.R
\name{add_season}
\alias{add_season}
\title{Add a season label based on outage day}
\usage{
add_season(df)
}
\arguments{
\item{df}{Data frame containing outage_day.}
}
\value{
Data frame with season factor.
}
\description{
Add a season label based on outage day
}
