% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingAnalysis.R
\name{missing_nonzero}
\alias{missing_nonzero}
\title{Return only variables with non-zero missingness}
\usage{
missing_nonzero(df, treat_ndas = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{treat_ndas}{Logical; if TRUE, converts "-NDA-" in XBORO to NA before summarising.}
}
\value{
A tibble filtered to na_count > 0.
}
\description{
Return only variables with non-zero missingness
}
