% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\name{package_data_path}
\alias{package_data_path}
\title{Path to example outage CSV}
\usage{
package_data_path()
}
\value{
A character string.
}
\description{
Returns the file path to the example outage data included with the package.
}
