% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventChains.R
\name{plot_chain_counts_by_boro}
\alias{plot_chain_counts_by_boro}
\title{Plot chain event counts by borough}
\usage{
plot_chain_counts_by_boro(chain_summary_df)
}
\arguments{
\item{chain_summary_df}{Output from summarise_event_chains().}
}
\value{
A ggplot object.
}
\description{
Plot chain event counts by borough
}
