% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventChains.R
\name{plot_chain_percent_by_boro}
\alias{plot_chain_percent_by_boro}
\title{Plot percent of outages in chains by borough}
\usage{
plot_chain_percent_by_boro(chain_summary_df)
}
\arguments{
\item{chain_summary_df}{Output from summarise_event_chains().}
}
\value{
A ggplot object.
}
\description{
Plot percent of outages in chains by borough
}
