% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventChains.R
\name{plot_chain_time_of_day}
\alias{plot_chain_time_of_day}
\title{Plot time-of-day distribution of chain outages (polar clock)}
\usage{
plot_chain_time_of_day(chains_df)
}
\arguments{
\item{chains_df}{Output from identify_event_chains().}
}
\value{
A ggplot object.
}
\description{
Plot time-of-day distribution of chain outages (polar clock)
}
