% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeeklyMonthlyTrend.R
\name{plot_customer_hours_weekly}
\alias{plot_customer_hours_weekly}
\title{Plot customer-hours lost per week}
\usage{
plot_customer_hours_weekly(weekly_df, start_week = NULL, end_week = NULL)
}
\arguments{
\item{weekly_df}{Weekly summary output from summarise_weekly_outages().}

\item{start_week}{Optional start week.}

\item{end_week}{Optional end week.}
}
\value{
A ggplot object.
}
\description{
Plot customer-hours lost per week
}
