% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DailySummary.R
\name{plot_daily_avg_restoration}
\alias{plot_daily_avg_restoration}
\title{Plot average restoration time (minutes) per day}
\usage{
plot_daily_avg_restoration(daily_df, start_date = NULL, end_date = NULL)
}
\arguments{
\item{daily_df}{A daily summary data frame.}

\item{start_date}{Optional start date.}

\item{end_date}{Optional end date.}
}
\value{
A ggplot object.
}
\description{
Plot average restoration time (minutes) per day
}
