% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingAnalysis.R
\name{plot_missing_bar}
\alias{plot_missing_bar}
\title{Plot missing counts by variable}
\usage{
plot_missing_bar(df, treat_ndas = TRUE, only_nonzero = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{treat_ndas}{Logical; if TRUE, converts "-NDA-" in XBORO to NA before summarising.}

\item{only_nonzero}{Logical; if TRUE, plots only variables with na_count > 0.}
}
\value{
A ggplot object.
}
\description{
Plot missing counts by variable
}
