% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingAnalysis.R
\name{plot_missing_upset}
\alias{plot_missing_upset}
\title{Plot missingness pattern via UpSet plot}
\usage{
plot_missing_upset(df, treat_ndas = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{treat_ndas}{Logical; if TRUE, converts "-NDA-" in XBORO to NA before plotting.}
}
\value{
A ggplot object.
}
\description{
Wrapper around naniar::gg_miss_upset().
}
