% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesDecomp.R
\name{plot_outage_trend_loess}
\alias{plot_outage_trend_loess}
\title{Plot daily outage trend using LOESS smoothing}
\usage{
plot_outage_trend_loess(daily_tsibble)
}
\arguments{
\item{daily_tsibble}{A tsibble returned by prepare_daily_ts().}
}
\value{
A ggplot object.
}
\description{
Plot daily outage trend using LOESS smoothing
}
