% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeeklyMonthlyTrend.R
\name{plot_outages_per_month}
\alias{plot_outages_per_month}
\title{Plot outages per month}
\usage{
plot_outages_per_month(monthly_df, start_month = NULL, end_month = NULL)
}
\arguments{
\item{monthly_df}{Monthly summary output from summarise_monthly_outages().}

\item{start_month}{Optional start month.}

\item{end_month}{Optional end month.}
}
\value{
A ggplot object.
}
\description{
Plot outages per month
}
