% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeeklyMonthlyTrend.R
\name{plot_seasonal_outage_frequency}
\alias{plot_seasonal_outage_frequency}
\title{Plot seasonal outage frequency}
\usage{
plot_seasonal_outage_frequency(df)
}
\arguments{
\item{df}{Data frame of outages.}
}
\value{
A ggplot object.
}
\description{
Plot seasonal outage frequency
}
