% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeeklyMonthlyTrend.R
\name{plot_troublecode_top_monthly}
\alias{plot_troublecode_top_monthly}
\title{Plot monthly outages by top trouble codes}
\usage{
plot_troublecode_top_monthly(df, top_n_codes = 5)
}
\arguments{
\item{df}{Data frame of outages.}

\item{top_n_codes}{Number of top trouble codes to include.}
}
\value{
A ggplot object.
}
\description{
Plot monthly outages by top trouble codes
}
