% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingAnalysis.R
\name{prep_missing_df}
\alias{prep_missing_df}
\title{Prepare data for missingness summaries}
\usage{
prep_missing_df(df, treat_ndas = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{treat_ndas}{Logical; if TRUE, converts "-NDA-" in XBORO to NA (if XBORO exists).}
}
\value{
A data frame.
}
\description{
Optionally converts placeholder "-NDA-" values in XBORO to NA.
}
