% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventChains.R
\name{summarise_event_chains}
\alias{summarise_event_chains}
\title{Summarise chain events by borough}
\usage{
summarise_event_chains(chains_df)
}
\arguments{
\item{chains_df}{Output from identify_event_chains().}
}
\value{
A tibble with total events, chain events, and percent chain events by borough.
}
\description{
Summarise chain events by borough
}
